fix card alpha when dismissing, update mock data params

+++ Tsunami/app/src/main/java/com/tonyjhuang/tsunami/ui/main/wave/contentview/WaveContentScrollView.java
@@ -136,7 +136,7 @@ public class WaveContentScrollView extends CardScrollView implements WaveContent
         }
 
         if(t < getCardViewStartingPosition()) {
-            setCardAlpha(t / getCardViewStartingPosition() + 0.2f);
+            setCardAlpha(((float) t) / getCardViewStartingPosition() + 0.2f);
         } else {
             setCardAlpha(1.0f);
         }
